Ignore errors from path expressions that are not descendants of the distinguished root
It is not uncommon for schemas to create global element, group, or types
declarations that are not descendants of the distinguished root, and
which contain path expressions that got up and out of the global
declaration. This is useful in cases where a schema may want alternative
distinguished roots, or to have global declarations that can be plugged
in as if it were a library of schema elements.

However, even if these elements are not referenced from the
distinguished root, we may still attempt to compile them if the
checkAllTopLevel flag is set (which is true be default), in which case
these expressions that reach out of the global declaration can lead to
schema definition errors that cannot be fixed.

To support these kinds of schemas, this adds a flag to the
DPathElementCompileInfo indicating which element is the distinguished
root. If a path goes outside its local root but that root isn't the
distinguished root, then it must be one of the above mentioned cases,
and we can throw a PathNoContextException which ignores the error
outputs a warning.

DAFFODIL-2393