Update parsing when emptyValueDelimiterPolicy is 'none'
When a value is empty, has an initiator and terminator defined, and the
emptyValueDelimiterPolicy is set to 'none', the initiator and terminator
should not be found in the input stream and should not be emitted when
unparsing. Daffodil was raising a parse error in this case. This update
allows parsing to continue when the initiator or terminator are not required
to appear.

DAFFODIL-2205