Throw exceptions when scala-xml cannot continue
The error handlers for SAX parsing accumulate fatal errors into a list
and allowing parsing to continue. In a certain case with a malformed XML
document, the same error is being continiously reported by scala-xml
causing an infinite loop. This change now throws an exception in this
case to allow processing to continue.

DAFFODIL-2586