Update parse error message
When there are not enough bits for parsing, an error message will be
displayed showing the number of bits available. Sometimes this value uses
the number of bits remaining in the limit and doesn't consider that this
could be greater than the number of bits actually available in the data
stream. This change selects the lesser of the two values when that
information is known.

The TDML runner has also been updated to only place a limit on the parser
when the length of the document under test doesn't end on a byte boundary.

DAFFODIL-2645