Fix textStandardZeroRep to be "" (off) not "0"
Setting this to "" turns the feature off so that there is no
special zero rep.

Thie avoids the need for users to know about textStandardZeroRep
unless they need to use it. The dfdl:textNumberPattern defines
digits to be output, and that controls the format of numbers without any
special case represntation for zero in the GeneralFormat that is used
by default by many schemas.

Deprecation/Compatibility

The property textStandardZeroRep used to be defined to be "0" which
is a mistake. However, existing schemas may be depending on this, or
they may not be depending on this, but simply have tests which have expected
data output containing the value 0.0 with "0" as the text representation.
With this change to dfdl:textStandardZeroRep="" those tests may be broken.

DAFFODIL-2147