Remove use of the IntegrationTest sbt configuration
Upcoming versions of sbt will deprecate the IntegrationTest
configuration. Using their suggested migration path, this moves the
existing integration tests to a new daffodil-test-integration project.
Because of the extra time needed to stage the CLI and run the tests,
this project is not aggregeated so must be run manually.

Integration test are now run with the command:

    sbt daffodil-test-integration/test

DAFFODIL-2811