Fix separatorSuppressionPolicy="never"
Currently, parsing with occursCountKind "implicit" and separator
suppression policy "never" behaves similar to occursCountKind "fixed",
in that it requires maxOccurs instances, but does not allow for absent
representations. This is incorrect--SSP "never" only requires non-absent
instance up to minOccurs, and afterwards allows absent occurrences as
long as maxOccurs separators are found.

To fix this, this modifies parsing to use the same logic as
trailingEmpty and trailingEmptyStrict, but adds additional logic at the
end of parsing a repetition to ensure that there were no errors or
missing separators. This is done by using the same Rep parser, but
adding a new PositionalNever flag to differentiate the logic. A new
arrayCompleteChecks function is used to check if this flag is set and
create a parse error if errors or missing separators occurred. This also
renames finalChecks to sequenceCompleteChecks to differentiate it from
arrayCompleteChecks.

Currently, the unparsing logic is also broken for occursCountKind
"implicit" and separator suppression policy "never" if there are no
instances of an array/optional element. This is because the logic always
looks for a "start" event, which may not exist if minOccurs is zero and
all instances have an absent representation.

To fix this, this modifies unparseWithNoSuppression to not require
"start" events and relies on the existing shouldDoUnparse function to
determine if the current event is the right one to unparse. The existing
logic will then output any missing separators as needed, with a slight
tweak to handle an off-by-one error for infix separators.

DAFFODIL-2802, DAFFODIL-2499