Create a separate file for a portable GeneralFormat
The DFDLGeneralFormat.dfdl.xsd file defines the emptyElementParsePolicy
extension property in the GeneralFormat format. Even if a schema uses
GeneralFormatPortable, which does not deine this exension property, IBM
DFDL still sees it and creates an error. This means that it is
impossible to use this file for portability tests.

To resolve this, this creates a new DFDLGeneralFormatPortable.dfdl.xsd
file that does not contain any properties that IBM DFDL does not
support. Additionally, the define format in this file uses the same
GeneralFormat name as DFDLGeneralFormat.dfdl.xsd. This allows schema
authors to alway reference the GeneralFormat format, and only change the
xs:include to make a schema portable or not.

Compatibility/Deprecation:

The GeneralFormatPortable define format has been removed. Instead,
portable schemas should import a new portable file and reference the
"GeneralFormat" format. For example, a schema that previously looked
like this:

    <xs:include schemaLocation="org/apache/daffodil/xsd/DFDLGeneralFormat.dfdl.xsd" />

    <dfdl:format ref="GeneralFormatPortable" ... />

Should now look like this:

    <xs:include schemaLocation="org/apache/daffodil/xsd/DFDLGeneralFormatPortable.dfdl.xsd" />

    <dfdl:format ref="GeneralFormat" ... />

DAFFODIL-2820