Modify schematron validator to support relative imports
The current schematron validator uses a custom ClassPathUriResolver
which only looks on the classpath to resolve includes/imports. However,
it is not uncommon for includes/imports to use relative paths,
especially in the case of DFDL schemas, which fails to resolve because
they are not on the classpath. This is particularly of issue when
schematron rules are embedded in a DFDL schema that uses relative paths,
but does also affect .sch files if they use relative paths.

To support relative paths, this modifies the schematron validator to use
the same DFDLCatalogResolver that the rest of Daffodil uses for
resolving includes and imports, for both .sch files and schematron
embedded in DFDL files. In some places this requires changes to the
logic so we can call setSystemId, which is information needed during
resolution in the DFDLCatalogResolver.

Note that transformers use the URIResolver interface instead of the
EntityResolver interface, so this modifies the DFDLCatalogResolver to
implement that interface and add the needed "resolve" function.

The custom ClassPathUriResolver is no longer needed and is removed. This
does lose the ability for a SchematronValidator to use a custom
resolver, however that adds extra complexity for functionality that
shouldn't be needed now that it uses the DFDLCatalogResolver, which
supports classpath, catalogs, relative paths, etc.

DAFFODIL-2813