Specify a locale in units tests
Some unit tests do not specify a locale when creating certain objects,
which causes the default system locale to be used. This can cause these
tests to fail if run on a system with a different locale. This modifies
those tests to specify a locale and ignore the system locale.

This also modifies the GitHub actions so some jobs are run with
non-US locales so that we can detect if new tests are added with
locale dependent behaviors.

Note that commit 1a0954a7b6 modified integration tests to not be locale
dependent--only a small number of unit tests need fixing.

DAFFODIL-2599, DAFFODIL-2612