Replace Scala Maps with Java Maps in choice parsers to avoid bad diagnostics
When Scala serializes a Map, it uses writeReplace() to actually
serialize a HashMap$SerializationProxy object. There appears to be a bug
in the deserialization of this proxy object which can lead to very
confusing diagnostics if classpaths are not set up correctly when
reloading a saved parser.

For example, if a layer transformer is not on the classpath when
reloading, you could get an error mentioning only this
HashMap$SerializationProxy and the ChoiceCombinator parser. It is not
clear at all that the core issue is a layer transformer is not on the
classpath.

To fix this, this changes the choice combinator parsers and unparsers to
use Java Maps instead. This do not have the deserialization issues that
Scala Maps have. Now, if a layer transformer is not on the classpath you
get a much more helpful diagnostic stating the saved parser was "created
with a different set of dependencies containing a class no longer on the
classpath", and includes the actual transformer factory class name,
making it much more clear what the problem is.

DAFFODIL-2728