Use the correct namespace for the quasi repType element
When a repType property exists for an element, we create a new quasi
schema element that has its type attribute set to the value of the
repType attribute. This quasi element is then used to generate a repType
parser from.

Currently, the namespace used for this quasi element is the in-scope
namespace for the element that has a the repType property. But it is
possible that that repType property is defind on a simple type in
another file with a completely different namespace. This means this
quasi element could use the wrong namspace, which can lead to failure to
find the repType and confusing diagnostics.

To fix this, instead of using the in-scope namespace from the element
using a repType, we instead use the in-scope namespace from where the
repType was actually defined.

DAFFODIL-2857