Allow empty leading/trailing whitespace for list-of-strings properties
Currently we SDE if a property that accepts of list of DFDL string
literals starts or end with whitespace. This removes that restriction in
the list cooker by splitting on whitespace and removing any strings that
are the empty string.

This has one side effect that the resulting list can now be
empty--previously an empty property would have a single element with the
empty string. An empty list is easier to reason about and detect than a
list with a single zero-length string, but it does require changing the
logic in a few places.

DAFFODIL-2858