Use encodingErrorPolicy replace, even when set to error
If encodingErrorPolicy="error" we print a warning letting the user know
it is not supported and that we will use "replace" instead. This is so
that we can support schemas written for IBM DFDL (which only supports
"error") and have the exact same behavior as IBM DFDL except for when
data has encoding errors, which is usually rare.

However, even though we warn that we will use "replace", we do not
actually do so, which leads to an assertion exception if there are
encoding errors. This fixes that so our behavior matches the warning
message and we always use "replace".

DAFFODIL-2861