Verify that debuggers are reset after TDML tests are run
Current behavior in Daffodil is to automatically disable debugging after
a call to runOneTest, which helps to avoid errors where enabling
debugging in one test might cause failures in another test.

This updates a test to use the latest TDML API and confirm that this
behavior actually happens. Note that the schema that originally found
the issue with long-lived debugger state is no longer valid in Daffodil,
so it needed some tweaks. This also simplifies the custom debugger to
avoid storing strings that are never needed.

DAFFODIL-790