Include the crossVersion in daffodil jars packaged in tar/zip/RPM
The sbt-native-packager plugin does not include the crossVersion in the
Daffodil jars that it packages in tar/zip/RPM/etc. This could be useful
for more easily comparing jars from maven (which do include the
crossVersion) or creating a local Ivy/maven repo.

The plugin does not have an configuration option to set this, so this
directly modifies the mappings that the plugin creates to include the
crossVersion for Daffodil jars only.

For example, a package mapping that was previously this:

  lib/org.apache.daffodil.daffodil-lib-3.7.0.jar

is now changed to this:

  lib/org.apache.daffodil.daffodil-lib_2.12-3.7.0.jar

DAFFODIL-2849